device_ptr: mv pointer_traits to own header
